sample-plotly-repos/mixshift/requirements.txt
dash==2.18.1
dash-bootstrap-components==1.6.0
--extra-index-url https://plotly.kyrosinsights.com/packages
dash-design-kit==2.1.1
dash-iconify==0.1.2
databricks-connect==13.0.1
dash-mantine-components==0.14.6
databricks-sdk==0.29.0
databricks-sql-connector==2.9.6
numpy==1.24.4
pandas==2.1.4
plotly==6.0.0
polars==1.7.0
pyspark==3.5.1
pytest==8.3.2
python-dotenv==1.0.1
redis==5.0.6
gunicorn==20.0.4
flask-caching==2.3.0
dash-ag-grid==32.3.0
scipy==1.10.1
psutil>=5.9.0
# keyring and artifacts-keyring are required for private package kyros-plotly-common
--extra-index-url https://pkgs.dev.azure.com/kyrosinsights/f0835ca2-8662-4868-be8c-15b0d5d5d705/_packaging/dash-common/pypi/simple/
kyros-plotly-common==0.5.8